+++ lib/lib/common/wrapper_dialog.dart
@@ -19,6 +19,8 @@ class WrapperDialog extends StatelessWidget {
     final borderRadius = timeState.widget.borderRadius ?? BORDER_RADIUS;
     final elevation = timeState.widget.elevation ?? ELEVATION;
 
+    debugPrint("width: ${timeState.widget.width}");
+
     return Dialog(
       insetPadding: timeState.widget.dialogInsetPadding,
       shape: RoundedRectangleBorder(
@@ -27,7 +29,11 @@ class WrapperDialog extends StatelessWidget {
       elevation: elevation,
       child: ClipRRect(
         borderRadius: BorderRadius.circular(borderRadius),
-        child: child,
+        child: SizedBox(
+          width: timeState.widget.width,
+          height: timeState.widget.height,
+          child: child,
+        ),
       ),
     );
   }